ingestion-requirements.txt
agate==1.9.1
aiohappyeyeballs==2.3.7
aiohttp==3.10.3
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cron-descriptor==1.4.3
daff==1.3.46
db-dtypes==1.3.0
dbt-adapters==1.4.1
dbt-bigquery==1.8.2
dbt-common==1.7.0
dbt-core==1.8.5
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
decorator==5.1.1
deepdiff==7.0.1
dlt==0.5.3
duckdb==0.10.3
frozenlist==1.4.1
fsspec==2024.6.1
gcsfs==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
giturlparse==0.12.0
google-api-core==2.19.1
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.10.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.5
grpcio-status==1.62.3
hexbytes==1.2.1
humanize==4.10.0
idna==3.7
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.4
jsonpath-ng==1.6.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
leather==0.4.0
Logbook==1.5.3
makefun==1.15.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mashumaro==3.13.1
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.4.0
msgpack==1.0.8
multidict==6.0.5
narwhals==1.4.1
networkx==3.3
numpy==2.0.1
oauthlib==3.2.2
ordered-set==4.1.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parsedatetime==2.6
pathspec==0.12.1
pathvalidate==3.2.0
pendulum==3.0.0
pillow==10.4.0
pipdeptree==2.9.6
ply==3.11
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requirements-parser==0.11.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
semver==3.0.2
shapely==2.0.6
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
streamlit==1.37.1
tenacity==8.5.0
text-unidecode==1.3
time-machine==2.15.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.1
types-setuptools==71.1.0.20240813
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
yarl==1.9.4
zipp==3.20.0
